Updated fonts to OpenDyslexic because accesibility.
Added subtitles Label to Dialog scene to show subtitles to non-spanish-readers.
Added icon that shows the player she can continue the dialog.
Created show animations for subtitles and continue icon.
Updated the environment color for the game and colors in fonts and DefaultButton scene.
Updated the animation to show the credits and the layout of main menu.
---- Ready to work on adding juice to the character frame ----