Added spritesheets for Walk and Talk animations in NPCs.
Updated NPC script so it changes states by default on certain events.
Updated the World's navigation path so Actors have more space for movement.
Fixed a bug in the Dialog flow that was preventing SPEAK nodes after MOVE nodes to trigger.
Added animation to CharacterFrame so it hides once an actor has spoken.
---- Ready to work on building two new scenes for Ep. 1 and 2. ----