Acting and tutorial sync.
Added overlay ColorRect for tutorials.
Updated the setup of the cameras in World.
Added navigation polygon for Stable.
World can now setup the set before "starting and episode".
When a dialog finishes, listeners can now the name of the finished dialog.
Rooms can return the absolute position of Points and Targets.
Updated dialogs to match the script defined by Itzel.
---- Ready to work on showing the letter and episode intros ----