Updated Mama Is expressions Sprite to have all of them.
Improved functionality of Autofill so it can change how the talking bubble renders depending on the character that speaks.
Adjusted camera zoom and positioning for both the Hacienda and the Stable.
Updated all the dialogs so they are better structured in the graph.
Updated expressions for Rico and Mama Is in episodes 3 and 5.
Created sprite for Mama Is talking bubble.